requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyPDF2==3.0.1
reportlab==4.0.7
google-generativeai==0.3.1
python-dotenv==1.0.0
